align(google): consolidate Google tests into GoogleIntegration/ section folder
