+++ tests/Humans.Application.Tests/GoogleIntegration/GoogleRemovalNotificationServiceTests.cs
@@ -10,7 +10,7 @@ using Microsoft.Extensions.Logging.Abstractions;
 using NSubstitute;
 using Xunit;
 
-namespace Humans.Application.Tests.Services;
+namespace Humans.Application.Tests.GoogleIntegration;
 
 /// <summary>
 /// Unit tests for <see cref="GoogleRemovalNotificationService"/> — the
