+++ tests/Humans.Application.Tests/GoogleIntegration/Infrastructure/StubGoogleGroupProvisioningClientTests.cs
@@ -4,7 +4,7 @@ using Humans.Infrastructure.Services.GoogleWorkspace;
 using Microsoft.Extensions.Logging.Abstractions;
 using Xunit;
 
-namespace Humans.Application.Tests.Infrastructure.GoogleWorkspace;
+namespace Humans.Application.Tests.GoogleIntegration.Infrastructure;
 
 /// <summary>
 /// Contract tests for <see cref="StubGoogleGroupProvisioningClient"/>. Pins
